chore(auth): refactor into domain layer
