+++ services/auth/src/domain/users.ts
@@ -1,0 +1,49 @@
+// User domain operations.
+//
+// The `user` table is owned primarily by better-auth (it does create/update
+// via OAuth callbacks + the admin plugin). We only need read access here for
+// the admin UI's user list. Mutations (set role, ban, delete) go through
+// better-auth's admin API directly from the handler — wrapping them here
+// would add an indirection hop with no value.
+import { drizzle } from 'drizzle-orm/d1'
+import { asc, desc } from 'drizzle-orm'
+import type { Env } from '../env'
+import { schema, user } from '../schema'
+
+const dbFor = (env: Env) => drizzle(env.AUTH_DB, { schema })
+
+export interface AdminUser {
+  id: string
+  email: string
+  name?: string
+  role?: string
+  banned?: boolean
+  banReason?: string | null
+}
+
+// Admins first (desc role string sorts 'admin' > 'cms-editor' > 'user'),
+// then alpha by email. Direct DB read rather than paginating through
+// better-auth's admin.listUsers — we already have the binding and this gives
+// us full control over ordering + the exact columns we need.
+export async function list(env: Env): Promise<AdminUser[]> {
+  const rows = await dbFor(env)
+    .select({
+      id: user.id,
+      email: user.email,
+      name: user.name,
+      role: user.role,
+      banned: user.banned,
+      banReason: user.banReason,
+    })
+    .from(user)
+    .orderBy(desc(user.role), asc(user.email))
+    .all()
+  return rows.map((r) => ({
+    id: r.id,
+    email: r.email,
+    name: r.name ?? undefined,
+    role: r.role ?? undefined,
+    banned: r.banned ?? undefined,
+    banReason: r.banReason ?? undefined,
+  }))
+}
